Add some likes/dislikes
I PR ironically, but seriously though: please write a programming language, I bet you could write a systems programming language that's ten times better than the archaic hell hole that is C.

There's so many cool code in the linux kernel that no developer ever sees because it's so hard to extract a module and release it as a package like is done in the scripting languages and for example Haskell.

With the right language you could write an operating system that really embeds in the developer community. Besides that I'd think taking design hints/shamelessly copying from Linux is fine. From a users perspective Linux is as good as perfect.

Please go for it, you're young and experienced, what more does a man need? ;)